Move gui to python bindings directory and remove run script
