+++ pyproject.toml
@@ -1,20 +1,23 @@
 [build-system]
-requires = ["scikit-build-core >=0.10", "nanobind >=1.3.2", "PyQt6 >=6.11.0" ]
+requires = ["scikit-build-core >=0.10", "nanobind >=1.3.2"]
 build-backend = "scikit_build_core.build"
 
 [project]
 name = "py-total-serialism"
-version = "0.0.1"
+version = "1.2.0"
 description = "An app that generates Total Serial sheet music"
 readme = "README.md"
 requires-python = ">=3.8"
 authors = [
-    { name = "Ryan Butler", email = "rmbultler@outlook.com" },
+    { name = "Ryan Butler", email = "rmbutler@outlook.com" }
 ]
 classifiers = [
     "License :: OSI Approved :: BSD License",
 ]
+dependencies = ["PyQt6 >=6.11.0"]
 
+[project.scripts]
+total-serialism = "py_total_serialism.cli:main"
 
 [tool.scikit-build]
 # Protect the configuration against future changes in scikit-build-core
