+++ python_bindings/py_total_serialism/consts.py
