+++ python_bindings/py_total_serialism/utils.py
