--- run.sh
@@ -1,2 +1,0 @@
-rm -rf __pycache__
-python3 python/main.py
